Open browser with maker UI automatically
There's no reason why we should constrain this behaviour to the taker only,
especially when it's useful when debugging/testing.
