+++ maker/src/main.rs
@@ -191,6 +191,7 @@ async fn main() -> Result<()> {
         .register("/", default_catchers())
         .attach(fairings::log_launch())
         .attach(fairings::log_requests())
+        .attach(fairings::ui_browser_launch())
         .launch()
         .await?;
