+++ taker/Cargo.toml
@@ -29,7 +29,6 @@ tokio-extras = { path = "../tokio-extras", features = ["xtra"] }
 tracing = { version = "0.1" }
 uuid = "0.8"
 vergen-version = { path = "../vergen-version" }
-webbrowser = "0.7.1"
 x25519-dalek = "1.1"
 xtra = { version = "0.6", features = ["instrumentation"] }
 xtra-bitmex-price-feed = { path = "../xtra-bitmex-price-feed" }
